Show Formatted JSON in the Staff Raw Books Data
Note that this uses code splitting so that it is only downloaded if the user is a moderator. At the price of some delay, we could probably further delay getting it to when someone actually expands the JSON panel, but it doesn't feel worth it; you only pay this price once then it will be cached.